Per #1597, update config file options to use dev_hit_radius, dev_hit_window, and opt_hit_tdiff. Also update log message to switch from 'lead' to 'forecast hour'.

+++ met/src/libcode/vx_tc_util/genesis_info.h
@@ -98,7 +98,7 @@ class GenesisInfo : public TrackInfo {
       double   dland()        const;
       unixtime genesis_time() const;
       int      genesis_lead() const;
-      double   genesis_fhr()  const;
+      int      genesis_fhr()  const;
 
          //
          //  do stuff
